waitm-449: refactor vitals table to optionaly display range info text
