+++ packages/mobile/App/ui/components/VitalsTable/index.tsx
@@ -1,7 +1,6 @@
-import React, { memo } from 'react';
+import React, { memo, useState } from 'react';
 import { PatientVitalsProps } from '../../interfaces/PatientVitalsProps';
 import { Table, TableCells } from '../Table';
-import { vitalsTableRows } from './VitalsTableRows';
 import { vitalsTableHeader } from './VitalsTableHeader';
 import { VitalsTableTitle } from './VitalsTableTitle';
 import { LoadingScreen } from '/components/LoadingScreen';
@@ -10,14 +9,28 @@ import { ErrorScreen } from '/components/ErrorScreen';
 import { VitalsDataElements } from '/helpers/constants';
 import { StyledText, StyledView } from '~/ui/styled/common';
 import { theme } from '~/ui/styled/theme';
+import { VitalsTableRowHeader } from './VitalsTableRowHeader';
+import { VitalsTableCell } from './VitalsTableCell';
+import { ValidationCriteria } from '~/types';
 
 interface VitalsTableProps {
   data: TableCells<PatientVitalsProps>;
   columns: string[];
 }
 
+const checkNeedsAttention = (
+  value: string,
+  validationCriteria: ValidationCriteria = {},
+) : boolean => {
+  const { normalRange } = validationCriteria;
+  const fValue = parseFloat(value);
+  if (!normalRange || Number.isNaN(fValue)) return false;
+  return fValue > normalRange.max || fValue < normalRange.min;
+};
+
 export const VitalsTable = memo(({ data, columns }: VitalsTableProps) : JSX.Element => {
   const [vitalsSurvey, error] = useBackendEffect(({ models }) => models.Survey.getVitalsSurvey());
+  const [showNeedsAttentionInfo, setShowNeedsAttentionInfo] = useState(false);
 
   if (!vitalsSurvey) {
     return <LoadingScreen />;
@@ -37,13 +50,33 @@ export const VitalsTable = memo(({ data, columns }: VitalsTableProps) : JSX.Elem
       <Table
         Title={VitalsTableTitle}
         tableHeader={vitalsTableHeader}
-        rows={vitalsTableRows(rows)}
+        rows={rows.map(r => {
+          const rowValidationCriteria = r.getValidationCriteriaObject();
+          return {
+            rowKey: 'dataElementId',
+            rowTitle: r.dataElementId,
+            rowHeader: () => <VitalsTableRowHeader title={r.dataElement.name} />,
+            cell: (cellData): JSX.Element => {
+              const needsAttention = checkNeedsAttention(cellData?.body || '', rowValidationCriteria);
+              if (needsAttention && !showNeedsAttentionInfo) setShowNeedsAttentionInfo(true);
+              return (
+                <VitalsTableCell
+                  data={cellData}
+                  key={cellData?.id || r.id}
+                  needsAttention={needsAttention}
+                />
+              );
+            },
+          };
+        })}
         columns={columns}
         cells={data}
       />
-      <StyledView padding={10}>
-        <StyledText color={theme.colors.ALERT}>*Vital needs attention</StyledText>
-      </StyledView>
+      {showNeedsAttentionInfo && (
+        <StyledView padding={10}>
+          <StyledText color={theme.colors.ALERT}>*Vital needs attention</StyledText>
+        </StyledView>
+      )}
     </>
   );
 });
